+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [[unpublished]](https://github.com/mlange-42/ark/compare/v0.8.0...main)
+
+### Bugfixes
+
+- Fixes error when removing multiple target entities of the same entity (#500)
+
 ## [[v0.8.0]](https://github.com/mlange-42/ark/compare/v0.7.1...v0.8.0)
 
 ### Features